Move vertexdecoder files into GPU/Common
