+++ GPU/GLES/TransformPipeline.h
@@ -21,7 +21,7 @@
 
 #include "GPU/Common/GPUDebugInterface.h"
 #include "GPU/Common/IndexGenerator.h"
-#include "GPU/GLES/VertexDecoder.h"
+#include "GPU/Common/VertexDecoderCommon.h"
 #include "gfx/gl_common.h"
 #include "gfx/gl_lost_manager.h"
