--- GPU/GPUXbox.vcxproj.filters
@@ -1,214 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup>
-    <Filter Include="GLES">
-      <UniqueIdentifier>{f7563dba-8146-4c21-a092-e864ff145d79}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Software">
-      <UniqueIdentifier>{4f6d1284-2c23-4ebc-842c-666a1305bfed}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Common">
-      <UniqueIdentifier>{21783292-4dd7-447b-af93-356cd2eaa4d6}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Null">
-      <UniqueIdentifier>{b31aa5a1-da08-47e6-9467-ab1d547b6ff3}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Directx9">
-      <UniqueIdentifier>{eb2a1d3d-24c7-4df8-b3cb-79a4b9734d70}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Directx9\helper">
-      <UniqueIdentifier>{862f23b4-2c1b-4d16-9450-caecbb77f276}</UniqueIdentifier>
-    </Filter>
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="ge_constants.h">
-      <Filter>Common</Filter>
-    </ClInclude>
-    <ClInclude Include="Math3D.h">
-      <Filter>Common</Filter>
-    </ClInclude>
-    <ClInclude Include="GLES\DisplayListInterpreter.h">
-      <Filter>GLES</Filter>
-    </ClInclude>
-    <ClInclude Include="GLES\FragmentShaderGenerator.h">
-      <Filter>GLES</Filter>
-    </ClInclude>
-    <ClInclude Include="GLES\Framebuffer.h">
-      <Filter>GLES</Filter>
-    </ClInclude>
-    <ClInclude Include="GLES\ShaderManager.h">
-      <Filter>GLES</Filter>
-    </ClInclude>
-    <ClInclude Include="GLES\TextureCache.h">
-      <Filter>GLES</Filter>
-    </ClInclude>
-    <ClInclude Include="GLES\TransformPipeline.h">
-      <Filter>GLES</Filter>
-    </ClInclude>
-    <ClInclude Include="GLES\VertexDecoder.h">
-      <Filter>GLES</Filter>
-    </ClInclude>
-    <ClInclude Include="GLES\VertexShaderGenerator.h">
-      <Filter>GLES</Filter>
-    </ClInclude>
-    <ClInclude Include="GPUState.h">
-      <Filter>Common</Filter>
-    </ClInclude>
-    <ClInclude Include="GPUInterface.h">
-      <Filter>Common</Filter>
-    </ClInclude>
-    <ClInclude Include="Null\NullGpu.h">
-      <Filter>Null</Filter>
-    </ClInclude>
-    <ClInclude Include="GLES\StateMapping.h">
-      <Filter>GLES</Filter>
-    </ClInclude>
-    <ClInclude Include="GLES\IndexGenerator.h">
-      <Filter>GLES</Filter>
-    </ClInclude>
-    <ClInclude Include="GeDisasm.h" />
-    <ClInclude Include="GPUCommon.h">
-      <Filter>Common</Filter>
-    </ClInclude>
-    <ClInclude Include="..\ext\xbrz\xbrz.h" />
-    <ClInclude Include="GLES\TextureScaler.h">
-      <Filter>GLES</Filter>
-    </ClInclude>
-    <ClInclude Include="Directx9\DisplayListInterpreter.h">
-      <Filter>Directx9</Filter>
-    </ClInclude>
-    <ClInclude Include="Directx9\FragmentShaderGenerator.h">
-      <Filter>Directx9</Filter>
-    </ClInclude>
-    <ClInclude Include="Directx9\Framebuffer.h">
-      <Filter>Directx9</Filter>
-    </ClInclude>
-    <ClInclude Include="Directx9\IndexGenerator.h">
-      <Filter>Directx9</Filter>
-    </ClInclude>
-    <ClInclude Include="Directx9\ShaderManager.h">
-      <Filter>Directx9</Filter>
-    </ClInclude>
-    <ClInclude Include="Directx9\StateMapping.h">
-      <Filter>Directx9</Filter>
-    </ClInclude>
-    <ClInclude Include="Directx9\TextureCache.h">
-      <Filter>Directx9</Filter>
-    </ClInclude>
-    <ClInclude Include="Directx9\TextureScaler.h">
-      <Filter>Directx9</Filter>
-    </ClInclude>
-    <ClInclude Include="Directx9\TransformPipeline.h">
-      <Filter>Directx9</Filter>
-    </ClInclude>
-    <ClInclude Include="Directx9\VertexDecoder.h">
-      <Filter>Directx9</Filter>
-    </ClInclude>
-    <ClInclude Include="Directx9\VertexShaderGenerator.h">
-      <Filter>Directx9</Filter>
-    </ClInclude>
-    <ClInclude Include="Directx9\helper\fbo.h">
-      <Filter>Directx9\helper</Filter>
-    </ClInclude>
-    <ClInclude Include="Directx9\helper\global.h">
-      <Filter>Directx9\helper</Filter>
-    </ClInclude>
-    <ClInclude Include="Directx9\helper\dx_state.h">
-      <Filter>Directx9\helper</Filter>
-    </ClInclude>
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="Math3D.cpp">
-      <Filter>Common</Filter>
-    </ClCompile>
-    <ClCompile Include="GLES\DisplayListInterpreter.cpp">
-      <Filter>GLES</Filter>
-    </ClCompile>
-    <ClCompile Include="GLES\FragmentShaderGenerator.cpp">
-      <Filter>GLES</Filter>
-    </ClCompile>
-    <ClCompile Include="GLES\Framebuffer.cpp">
-      <Filter>GLES</Filter>
-    </ClCompile>
-    <ClCompile Include="GLES\ShaderManager.cpp">
-      <Filter>GLES</Filter>
-    </ClCompile>
-    <ClCompile Include="GLES\TextureCache.cpp">
-      <Filter>GLES</Filter>
-    </ClCompile>
-    <ClCompile Include="GLES\TransformPipeline.cpp">
-      <Filter>GLES</Filter>
-    </ClCompile>
-    <ClCompile Include="GLES\VertexDecoder.cpp">
-      <Filter>GLES</Filter>
-    </ClCompile>
-    <ClCompile Include="GLES\VertexShaderGenerator.cpp">
-      <Filter>GLES</Filter>
-    </ClCompile>
-    <ClCompile Include="GPUState.cpp">
-      <Filter>Common</Filter>
-    </ClCompile>
-    <ClCompile Include="Null\NullGpu.cpp">
-      <Filter>Null</Filter>
-    </ClCompile>
-    <ClCompile Include="GLES\StateMapping.cpp">
-      <Filter>GLES</Filter>
-    </ClCompile>
-    <ClCompile Include="GLES\IndexGenerator.cpp">
-      <Filter>GLES</Filter>
-    </ClCompile>
-    <ClCompile Include="GeDisasm.cpp" />
-    <ClCompile Include="GPUCommon.cpp">
-      <Filter>Common</Filter>
-    </ClCompile>
-    <ClCompile Include="..\ext\xbrz\xbrz.cpp" />
-    <ClCompile Include="GLES\TextureScaler.cpp">
-      <Filter>GLES</Filter>
-    </ClCompile>
-    <ClCompile Include="Directx9\DisplayListInterpreter.cpp">
-      <Filter>Directx9</Filter>
-    </ClCompile>
-    <ClCompile Include="Directx9\FragmentShaderGenerator.cpp">
-      <Filter>Directx9</Filter>
-    </ClCompile>
-    <ClCompile Include="Directx9\Framebuffer.cpp">
-      <Filter>Directx9</Filter>
-    </ClCompile>
-    <ClCompile Include="Directx9\IndexGenerator.cpp">
-      <Filter>Directx9</Filter>
-    </ClCompile>
-    <ClCompile Include="Directx9\ShaderManager.cpp">
-      <Filter>Directx9</Filter>
-    </ClCompile>
-    <ClCompile Include="Directx9\StateMapping.cpp">
-      <Filter>Directx9</Filter>
-    </ClCompile>
-    <ClCompile Include="Directx9\TextureCache.cpp">
-      <Filter>Directx9</Filter>
-    </ClCompile>
-    <ClCompile Include="Directx9\TextureScaler.cpp">
-      <Filter>Directx9</Filter>
-    </ClCompile>
-    <ClCompile Include="Directx9\TransformPipeline.cpp">
-      <Filter>Directx9</Filter>
-    </ClCompile>
-    <ClCompile Include="Directx9\VertexDecoder.cpp">
-      <Filter>Directx9</Filter>
-    </ClCompile>
-    <ClCompile Include="Directx9\VertexShaderGenerator.cpp">
-      <Filter>Directx9</Filter>
-    </ClCompile>
-    <ClCompile Include="Directx9\helper\fbo.cpp">
-      <Filter>Directx9\helper</Filter>
-    </ClCompile>
-    <ClCompile Include="Directx9\helper\dx_state.cpp">
-      <Filter>Directx9\helper</Filter>
-    </ClCompile>
-    <ClCompile Include="Directx9\helper\global.cpp">
-      <Filter>Directx9\helper</Filter>
-    </ClCompile>
-    <ClCompile Include="Directx9\Spline.cpp">
-      <Filter>Directx9</Filter>
-    </ClCompile>
-  </ItemGroup>
-</Project>
